test: relocate LBA unit tests
